refactor: Expose scale factor for parameter errors through ICostFunction interface
This allows to define cost functions where parameter errors scale
differently than in least squares.
Parameter errors are defined as the change in parameter value required
to change the function value by the value of the new property `Up`.
Normally, for chi-squared fits Up = 1 (for negative log likelihood it
would be 0.5). If the user wants instead 2-sigma parameter errors,
for chi-squared fits, the value would need to be set to 4. See the
Minuit2 docs for more detail.