refactor: Revise assertion of number of function calls
Instead of testing whether they have surpassed a threshold (which never
really made sense), we now compare if they are close to the iminuit
(reference) results, but only for the cases where a comparison is
possible/straight-forward. For example, we drop the assert for the
test cases with missing y-errors.