fix: Fix lower parameter limit being ignored when upper limit is set
This is a pretty nasty bug (or, at least, usability issue) in Minuit2:
When using the SetUpperLimit() method on MnUserParameterState, the
previously set lower limit is ignored, or rather the parameter
projection strategy is to "fUpperLimTrafo", instead of
"fDoubleLimTrafo". The SetLimits() method leads to the expected result
in the presence of both limits.

The test defining the behavior in case of covariance calculation
failure had to be dropped since the bug fix circumvents the covariance
failure for the test case, and I did not manage to find another
suitable test case. The graceful handling of covariance failures is
left in the production code nevertheless, since failures may still
occur.