feat: Revise creation of parameter configuration
The static constructors for fixed (constant) and variable parameters
avoid unwanted interference between fixed and limited parameters.
To my mind, a fixed parameter / constant should not have (or be
concerned by its) limits, as they have no effect for the minimization
anyways. If the user wants to avoid unphysical values for constants, he
or she should handle this upfront.