refactor: Rename Precise strategy to Rigorous to improve clarity
The term Precise could be interpreted as if the minimization result
would be more precise. This is not necessarily true, and not the intend
of the setting. Rigorous is considered to be a better term, since what
actually happens is that the minimizer wastes function calls
(sacrifices speed) in order to calculate the matrix of second
derivatives explicitly to converge more reliably.

See the Miunuit documentation [1] and the more detailed iminuit FAQ
section [2] for further info.

[1] https://root.cern.ch/doc/master/Minuit2Page.html
[2] https://github.com/scikit-hep/iminuit/blob/develop/doc/faq.rst