feat: Remove automated Hesse execution for cost function requiring error definition adjustment from MigradMinimizer
Hereafter, it is the users choice if they want to do refine the
parameter errors after the minimization in this scenario, or if they
don't need valid parameter errors (and prefer a faster result), or if
they even want to continue differently, e.g. using another minimization
cycle after error definition adjustment.