feat: Ignore excess parameter configurations passed to minimization methods
Before, parameter configurations containing excess configurations irrelevant to
the cost function were leading to an exception. This was not really necessary,
though, when all configurations required by the cost function were uniquely
contained.
Hereafter, excess configurations are simply ignored. This simplifies working
with the tool, since the same set of parameter configurations can now be reused
to minimize cost functions with related models. For instance, a set of parameter
configurations defining "slope" and "offset" can now be used for a line model
with and without an "offset" parameter.

BREAKING-CHANGE: Changes behavior.