refactor: Improve cost function monitoring
So far, the last valid parameter values passed to the cost function were
monitored. The problem is that those values could not only represent the
(wanted) last valid state, but could also the last values used to numerically
approximate the gradient. In the latter case, those values could be worse than
the initial parameter values.
Hereafter, the monitor records the last best values, i.e. the parameter values
representing the minimal cost value up to the point of exit.