feat: Revise handling of cancelling and non-finite values to make CostFunctionAdapter thread-safe
This is to prepare for multi-threading. The adapter no longer has mutable state,
but forwards its last input through exceptions.
Hereafter, "premature minimization results" all represent the last state of the
minimization at which an exception (cancellation of non-finite value) occurred.
It no longer represents the last VALID state.
Because the number of fauntion calls is no longer tracked, the results expose a
null for this property for premature terminations.