refactor: Adjust test such that they reliably succeed for enabled multi-threading
The issue is that with enabled OpenMP there is a certain degree of
non-deterministic behavior with respect to function call order.
Especially, number of function calls are not strictly reproducible.
This is a known property of Minuit2 with enabled OpenMP. Test,
function call tolerances in particular, were adjusted to still
succeed reliably.