ci: Fix packing step in ci pipeline
The dotnet pack step was re-building the project but not listening to
the UseOpenMP flag, meaning the "*.openmp package" was not actually
using OpenMP support. Since everything is built beforehand, it should
be safe and more efficient to use the --no-build flag for the packing
step. The --disable-build-servers flag should be redunant after this
change and is hence removed.