refactor: Skip assertions of number of function calls when OpenMP support is enabled
It is known that the number of function calls differs and is not
fully deterministic when OpenMP support is active. (see
https://indico.cern.ch/event/36763/contributions/871192/attachments
/730869/1002804/ParallelMinuit_27Jun.pdf)

This was leading to annoying occasional failures of the associated
tests in the CI/CD pipeline even for large tolerances. Increasing the
tolerances further does not really make sense.