refactor: Extend exception handling across interop boundary
I found that for the Hessian it is not possible to trigger process
termination by (non-finite) signal values. Hence, we needed another
way for process termination. I found that only the cost value
method is called concurrently and has issues with immediate abort
via a c++ exception. Gradient and Hessian are not affected and can
be terminated immediately via a c++ exception.

Adding a success indicator to the c++ process results indicating
termination caused by exception has the positive side effect of
handling also all other potential exceptions that might occur on
the c++ side.

In the course of this commit, all constructor and method signatures
in the custom c++ wrappers are consistently changed to clang-format:
const T& name).