refactor: Fix forwarding of HasHessian and HasHessianDiagonal in ComponentCostFunction
This error was not having any effect, since the CostFunctionSum was
calculating its HasHessian and HasHessianDiagonal properties from the
original (non-wrapped) components. The proper forwarding through the
ComponentCostFunction is still cleaner since it avoids unexpected
errors if we ever decide to wrap the components before extracting
their HasHessian and HasHessianDiagonal properties.