fix: Fix occasional memory corruption when trying to refine minimization result
C# was taking ownership of the newly introduced RunResult, but not of
its wrapped FunctionMinimum. Consequently, when the short-lived
RunResult when out of scope at the end of the `Minimize` method, the
native FunctionMinimum was freed, which could happen fast enough such
that the subsequent `HesseErrorCalculator.Refine` when trying to
access the FunctionMinimum was trying to access the freed memory,
which led to a SystemAccessViolation error.
An additional copy step now guarantees that C# takes owenership of
the FunctionMinimum that is kept with the MinimizationResult, too.